% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexgrid.r
\name{create_hexgrid}
\alias{create_hexgrid}
\title{Make a uniform hexgrid version of a GADM Germany shapefile}
\usage{
create_hexgrid()
}
\value{
a SpatialPolygons hex grid
}
\description{
This reads in an Admin 0 shapefile of Germany using \code{getData}
then turns that into a hexgrid map using built-in spatial routines.
}

