% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zellingenach.r
\name{display_maps}
\alias{display_maps}
\title{Call this function to make magic happen!}
\usage{
display_maps(output_file = NULL)
}
\arguments{
\item{output_file}{where to save the built HTML (optional)}
}
\description{
This is the function that makes the HTML page with the maps.
}
\details{
It calls \code{make_maps()} if it hasn't been called before and
caches the resultant data, then builds the HTML page.
}

