% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zellingenach.r
\name{make_maps}
\alias{make_maps}
\title{Do all the hard work}
\usage{
make_maps(verbose = TRUE)
}
\arguments{
\item{verbose}{tell folks what's going on}
}
\value{
a \code{list} with an element for each suffix group that
  contains the title, subtitle, ggplot2 object and count of towns.
}
\description{
This should be called to build the data structures for the ultimate
map / HTML page production.
}
\details{
This function reads in the places data, builds the base map hexgrid,
identifies which towns belong in which hex, assigns the colors
(assignment is based on a log scale of the colors vs an pre-determined
cutoff in the pure-javascript version) then builds a \code{list}
with an element for each suffix group that contains the title,
subtitle, ggplot2 object and count of towns.
}

