% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/places.r
\name{read_places}
\alias{read_places}
\title{Retrieve town names and determine which suffix bucket they belong in}
\usage{
read_places(suf = suffix_regex())
}
\arguments{
\item{suf}{list of regular expressions to match suffixes}
}
\value{
\code{data.frame} of enriched places
}
\description{
The CSV file that Mortiz used is bundled with this package. You can find
it at \code{system.file("extdata/placenames_de.tsv", package="zellingenach")}
}
\details{
It reads the names, determines which suffix group a town belongs in and
then filters out all the ones that weren't found.
}
\note{
My version is more inclusive than Moritz's (if a town matches more
      than one suffix it will be counted in more than one suffix group).
}

