% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/suffixes.r
\name{suffix_regex}
\alias{suffix_regex}
\title{Retrieve suffix list as end-matching regexes}
\usage{
suffix_regex()
}
\value{
A character vector of regular expressions
}
\description{
A slightly modified version of the suffix group list from Moritz's post
is included with this package. You can find it at
\code{system.file("extdata/suffixlist.js", package="zellingenach")} and it is
read in using the \code{V8} package since it was left as a javascript
data structure.
}

