% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zellingenach-package.r
\docType{package}
\name{zellingenach}
\alias{zellingenach}
\alias{zellingenach-package}
\title{-zell, -ingen, -ach}
\description{
A package to (mostly) replicate Moritz Stefaner's inaugural 2016 post
\url{http://truth-and-beauty.net/experiments/ach-ingen-zell/}
}
\details{
To see the maps, call \code{display_maps()}

Moritz's inaugural 2016 post is a visual exploration of the spatial
patterns in the endings of German town and village names.

He picked the most interesting suffixes from
\url{https://de.wikipedia.org/wiki/Ortsname} and cross-referenced them with
a list of place names from \href{http://www.geonames.org/export/}{geonames}.
(Note: The approach is not 100%% scientific, as he only match the letters at
the end of the string, not actual syllables.)

He did his in javascript (\url{http://github.com/moritzstefaner/ach-ingen-zell}).
}
\author{
Bob Rudis (@hrbrmstr)
}

