% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffixes.r
\name{suffix_names}
\alias{suffix_names}
\title{Retrieve suffix list as a list of vectors with suffix names}
\usage{
suffix_names()
}
\value{
A list of suffix character vectors
}
\description{
A slightly modified version of the suffix group list from Moritz's post
is included with this package. You can find it at
\code{system.file("alt/suffixlist.js", package="zellingenach")} and it is
read in using the \code{V8} package since it was left as a javascript
data structure.
}

